Move PageHandler and ViewTitle to layouts/components
